[Refactor] Remove config.h from headers P2
