+++ be/src/exec/es/es_scroll_parser.cpp
@@ -20,7 +20,6 @@
 #include "column/array_column.h"
 #include "column/column_helper.h"
 #include "column/nullable_column.h"
-#include "common/config.h"
 #include "runtime/runtime_state.h"
 #include "types/logical_type.h"
 #include "types/timestamp_value.h"
